+++ src/components/CollectionItemCard/DownloadLink.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import { Icon } from "@nypl/design-system-react-components";
+import { Button, Icon } from "@nypl/design-system-react-components";
 import CollectionUtils from "~/src/util/CollectionUtils";
 import Link from "~/src/components/Link/Link";
 import { OpdsLink } from "~/src/types/OpdsModel";
-import * as gtag from "~/src/lib/Analytics";
 import { formatUrl } from "~/src/util/Util";
+import { trackEvent } from "~/src/lib/Analytics";
 
 const DownloadLink: React.FC<{ links: OpdsLink[]; title: string }> = ({
   links,
@@ -14,23 +14,35 @@ const DownloadLink: React.FC<{ links: OpdsLink[]; title: string }> = ({
 
   if (!selectedLink) return null;
 
+  const formattedUrl = formatUrl(selectedLink.href);
+
+  const trackDownloadCta = () => {
+    trackEvent({
+      section: `${title}`,
+      text: "Download",
+      destination: `${formattedUrl}`,
+    });
+  };
+
   if (selectedLink && selectedLink.href) {
     return (
-      <Link
-        to={`${formatUrl(selectedLink.href)}`}
-        linkType="action"
-        onClick={() => {
-          gtag.drbEvents("Download", `${title}`);
-        }}
-      >
-        <Icon
-          name="download"
-          align="left"
-          size="small"
-          decorative
-          iconRotation="rotate0"
-        />
-        Download PDF
+      <Link to={`${formattedUrl}`} linkType="action">
+        <Button
+          id="download-button"
+          buttonType="secondary"
+          onClick={() => {
+            trackDownloadCta();
+          }}
+        >
+          <Icon
+            name="download"
+            align="left"
+            size="small"
+            decorative
+            iconRotation="rotate0"
+          />
+          Download PDF
+        </Button>
       </Link>
     );
   }