+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Update to Web Reader version 4.3.1
 - Add TOC for single resource PDFs
 - Refactored Playwright/Cucumber tests to use new dropdown action function
+- Migrate from Google Analytics to Adobe Analytics
 
 ## [0.17.1]
 - Replace dgx-header-component and DS Footer with Header/Footer components from nypl-header-app